Second test kept failing
The constant baseAccount on line 11 was not falling out of scope for the second test case so I moved it outside the test and made it a global variable. The test cases now run without failure. 

Error I was getting: 
`TypeError: Assignment to constant variable.`
` ReferenceError: Cannot access 'baseAccount' before initialization`